Fix CLI import path
