+++ setup.py
@@ -3,17 +3,20 @@
 Setup script for Fun with Maps.
 """
 
-from setuptools import setup, find_packages
 import os
+
+from setuptools import find_packages, setup
 
 # Read the contents of your README file
 this_directory = os.path.abspath(os.path.dirname(__file__))
-with open(os.path.join(this_directory, 'README.md'), encoding='utf-8') as f:
+with open(os.path.join(this_directory, "README.md"), encoding="utf-8") as f:
     long_description = f.read()
 
 # Read requirements from requirements.txt
-with open(os.path.join(this_directory, 'requirements.txt'), encoding='utf-8') as f:
-    requirements = [line.strip() for line in f if line.strip() and not line.startswith('#')]
+with open(os.path.join(this_directory, "requirements.txt"), encoding="utf-8") as f:
+    requirements = [
+        line.strip() for line in f if line.strip() and not line.startswith("#")
+    ]
 
 setup(
     name="fun-with-maps",
@@ -54,9 +57,9 @@ setup(
     entry_points={
         "console_scripts": [
             "fun-with-maps=scripts.main:main",
-            "fun-with-maps-cli=scripts.cli:cli",
+            "fun-with-maps-cli=fun_with_maps.cli:cli",
         ],
     },
     include_package_data=True,
     zip_safe=False,
-) 
+)
